refactor: controller refactoring
