+++ SRC/BackEnd_API/src/main/java/com/ringtaillemur/rainmaker/controller/api/dorametric/DeploymentFrequencyController.java
@@ -1,0 +1,31 @@
+package com.ringtaillemur.rainmaker.controller.api.dorametric;
+
+import java.time.LocalDate;
+import java.util.List;
+
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ringtaillemur.rainmaker.dto.webdto.responsedto.DeploymentFrequencyDto;
+import com.ringtaillemur.rainmaker.service.dorametrics.DeploymentFrequencyService;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController("/api/dorametric/deployment-frequency")
+@RequiredArgsConstructor
+public class DeploymentFrequencyController {
+
+	private final DeploymentFrequencyService deploymentFrequencyService;
+
+	@GetMapping
+	public DeploymentFrequencyDto doraDeploymentFrequencyForRestApi(
+		@RequestParam("start_time") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate startTime,
+		@RequestParam("end_time") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate endTime,
+		@RequestParam("repo_id") List<Long> repositoryIds) {
+		return deploymentFrequencyService.getDeploymentFrequency(repositoryIds, startTime, endTime);
+	}
+
+}
